+++ src/Shared/NcdfUtil/hco_m_netcdf_io_define.F90
@@ -89,11 +89,8 @@ CONTAINS
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT NONE
-!
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !!  ncid  : netCDF file id
@@ -130,14 +127,14 @@ CONTAINS
     len0 = len
     if (present(unlimited)) then
        if (unlimited) then
-          len0 = NF_UNLIMITED
+          len0 = NF90_UNLIMITED
        endif
     endif
 
-    ierr = Nf_Def_Dim (ncid, name, len0, id)
+    ierr = NF90_Def_Dim(ncid, name, len0, id)
 
-    IF (ierr.ne.NF_NOERR) then
-       err_msg = 'Nf_Def_Dim: can not define dimension : '// Trim (name)
+    IF (ierr.ne.NF90_NOERR) then
+       err_msg = 'NF90_Def_Dim: can not define dimension : '// Trim (name)
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
     END IF
 
@@ -153,33 +150,31 @@ CONTAINS
 !
 ! !INTERFACE:
 !
-  SUBROUTINE NcDef_variable(ncid,name,type,ndims,dims,var_id,compress)
+  SUBROUTINE NcDef_variable(ncid, name, xtype, ndims, dims, var_id, compress)
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT NONE
-!
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !
 !!  ncid   : netCDF file id
 !!  name   : name of the variable
 !!  type   : type of the variable
-!!           (NF_FLOAT, NF_CHAR, NF_INT, NF_DOUBLE, NF_BYTE, NF_SHORT)
+!!           (NF90_FLOAT,  NF90_CHAR, NF90_INT,
+!!            NF90_DOUBLE, NF90_BYTE, NF90_SHORT)
 !!  ndims  : number of dimensions of the variable
 !!  dims   : netCDF dimension id of the variable
     CHARACTER (LEN=*), INTENT(IN)  :: name
     INTEGER,           INTENT(IN)  :: ncid, ndims
     INTEGER,           INTENT(IN)  :: dims(ndims)
-    INTEGER,           INTENT(IN)  :: type
+    INTEGER,           INTENT(IN)  :: xtype
     LOGICAL, OPTIONAL, INTENT(IN)  :: compress
 !
 ! !OUTPUT PARAMETERS:
 !
-!!  varid  : netCDF variable id returned by NF_DEF_VAR
+!!  varid  : netCDF variable id returned by NF90_DEF_VAR
     INTEGER,           INTENT(OUT) :: var_id
 !
 ! !DESCRIPTION: Defines a netCDF variable.
@@ -195,59 +190,67 @@ CONTAINS
 !BOC
 !
 ! !LOCAL VARIABLES:
-    character (len=512) :: err_msg
-    integer ::  ierr
-    logical ::  doStop
-    ! Compression settings
-    ! choose deflate_level=1 for fast, minimal compression.
-    ! Informal testing suggests minimal benefit from higher compression level
-    integer, parameter :: shuffle=1, deflate=1, deflate_level=1
-!
-    ierr = Nf_Def_Var (ncid, name, type, ndims, dims, var_id)
+    character(len=512) :: err_msg
+    integer            :: ierr
+    logical            :: doStop
 
-    IF (ierr.ne.NF_NOERR) THEN
-       err_msg = 'Nf_Def_Var: can not define variable : '// Trim (name)
-       CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
-    END IF
-
-#if defined( NC_HAS_COMPRESSION )
+#ifdef NC_HAS_COMPRESSION
     !=====================================================================
-    ! If the optional "compress" variable is used and set to TRUE,
-    ! then enable variable compression (cdh, 0/17/17)
+    ! Create a compressed (deflated) netCDF variable
     !
     ! NOTE: We need to block this out with an #ifdef because some
-    ! netCDF installations might lack the nf_def_var_deflate function
+    ! netCDF installations might lack the deflation capability,
     ! which would cause a compile-time error. (bmy, 3/1/17)
-    !
-    ! ALSO NOTE: Newer versions of netCDF balk when you try to compress
-    ! a scalar variable.  This generates an annoying warning message.
-    ! To avoid this, only compress array variables. (bmy, 11/30/20)
-    !=====================================================================
-    if (present(Compress) .and. ndims > 0) then
+    !========================================================================
+    IF ( PRESENT( Compress ) ) then
 
-       if (Compress) then
+       !------------------------------------------------------------------
+       ! If COMPRESS is passed as an optional argument, and is TRUE,
+       ! then define the variable with deflate_level=1.  Higher values
+       ! of deflate_level yield minimal additiional benefit.
+       !
+       ! ALSO NOTE: Newer versions of netCDF balk when you try to compress
+       ! a scalar variable.  This generates an annoying warning message.
+       ! To avoid this, only compress array variables. (bmy, 11/30/20)
+       !-------------------------------------------------------------------
+       IF ( Compress .and. ndims > 0 ) THEN
 
-          ! Set compression
-          ierr = nf_def_var_deflate( ncid, var_id,  shuffle,       &
-                                           deflate, deflate_level )
+          ! Create deflated variable
+          ierr = NF90_Def_Var( ncid, name, xtype, dims, var_id,              &
+                               shuffle=.TRUE., deflate_level=1 )
 
           ! Check for errors.
           ! No message will be generated if the error is simply that the
-          ! file is not netCDF-4
-          ! (i.e. netCDF-3 don't support compression)
-          IF ( (ierr.ne.NF_NOERR) .and. (ierr.ne.NF_ENOTNC4)) THEN
+          ! file is not netCDF-4 (as netCDF-3 doesn't support compression)
+          IF ( (ierr.ne.NF90_NOERR) .and. (ierr.ne.NF90_ENOTNC4)) THEN
 
              ! Errors enabling compression will not halt the program
              doStop = .False.
 
              ! Print error
-             err_msg = 'Nf_Def_Var_Deflate: can not compress variable : '// Trim (name)
+             err_msg = 'NF90_Def_Var: can not create compressed variable : '//&
+                        Trim(name)
              CALL Do_Err_Out (err_msg, doStop, 0, 0, 0, 0, 0.0d0, 0.0d0)
           END IF
 
-       endif
-    endif
+          ! Return successfully
+          RETURN
+       ENDIF
+    ENDIF
 #endif
+
+    !========================================================================
+    ! Create an uncompressed netCDF variable if:
+    ! (1) COMPRESS is not passed as an optional argument
+    ! (2) COMPRESS is passed as an optional argument but is FALSE
+    ! (3) The variable is a scalar (ndims == 0)
+    !========================================================================
+    ierr = NF90_Def_Var( ncid, name, xtype, dims, var_id )
+    IF ( ierr /= NF90_NOERR ) THEN
+       err_msg = 'NF90_Def_Var_Deflate: can not create variable : '// &
+            Trim (name)
+       CALL Do_Err_Out (err_msg, doStop, 0, 0, 0, 0, 0.0d0, 0.0d0)
+    ENDIF
 
   END SUBROUTINE NcDef_variable
 !EOC
@@ -261,14 +264,12 @@ CONTAINS
 !
 ! !INTERFACE:
 !
-  SUBROUTINE NcDef_var_attributes_c(ncid,var_id,att_name,att_val)
+  SUBROUTINE NcDef_var_attributes_c(ncid, var_id, att_name, att_val)
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT none
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !!  ncid    : netCDF file id
@@ -276,7 +277,7 @@ CONTAINS
 !!  att_name: attribute name
 !!  att_val : attribute value
     CHARACTER (LEN=*), INTENT(IN) :: att_name, att_val
-    INTEGER,           INTENT(IN) :: ncid, var_id
+    INTEGER,           INTENT(IN) :: ncid,     var_id
 !
 ! !DESCRIPTION: Defines a netCDF variable attribute of type: CHARACTER.
 !\\
@@ -292,12 +293,11 @@ CONTAINS
 !
 ! !LOCAL VARIABLES:
     CHARACTER (LEN=512) :: err_msg
-    INTEGER             :: mylen, ierr
+    INTEGER             :: ierr
 !
-    mylen = LEN(att_val)
-    ierr = Nf_Put_Att_Text (ncid, var_id, att_name, mylen, att_val)
+    ierr = NF90_Put_Att(ncid, var_id, att_name, att_val)
 
-    IF (ierr.ne.NF_NOERR) THEN
+    IF (ierr /= NF90_NOERR) THEN
        err_msg = 'NcDef_var_attributes_c: can not define attribute : ' // &
             TRIM (att_name)
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
@@ -315,10 +315,11 @@ CONTAINS
 !
 ! !INTERFACE:
 !
-  SUBROUTINE NcDef_var_attributes_i(ncid,var_id,att_name,att_val)
+  SUBROUTINE NcDef_var_attributes_i(ncid, var_id, att_name, att_val)
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
 !
     IMPLICIT NONE
@@ -350,11 +351,9 @@ CONTAINS
     character (len=512) :: err_msg
     integer             :: mylen, ierr
 !
-    mylen = 1
-    ierr  = Nf_Put_Att_Int( ncid,   var_id, att_name, &
-                            NF_INT, mylen,  att_val )
+    ierr  = NF90_Put_Att( ncid, var_id, att_name, att_val )
 
-    IF (ierr.ne.NF_NOERR) THEN
+    IF (ierr.ne.NF90_NOERR) THEN
        err_msg = 'NcDef_var_attributes_i: can not define attribute : ' // &
             TRIM (att_name)
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
@@ -372,14 +371,12 @@ CONTAINS
 !
 ! !INTERFACE:
 !
-  SUBROUTINE NcDef_var_attributes_r4(ncid,var_id,att_name,att_val)
+  SUBROUTINE NcDef_var_attributes_r4(ncid, var_id, att_name, att_val)
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-
-    IMPLICIT NONE
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !!  ncid    : netCDF file id
@@ -405,13 +402,11 @@ CONTAINS
 !
 ! !LOCAL VARIABLES:
     CHARACTER (LEN=512) :: err_msg
-    INTEGER             :: mylen, ierr
+    INTEGER             :: ierr
 !
-    mylen = 1
-    ierr  = Nf_Put_Att_Real( ncid,     var_id, att_name, &
-                             NF_FLOAT, mylen,  att_val )
+    ierr  = NF90_Put_Att( ncid, var_id, att_name, att_val )
 
-    IF (ierr.ne.NF_NOERR) THEN
+    IF (ierr.ne.NF90_NOERR) THEN
        err_msg = 'NcDef_var_attributes_r4: can not define attribute : ' // &
             TRIM (att_name)
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
@@ -429,14 +424,12 @@ CONTAINS
 !
 ! !INTERFACE:
 !
-  SUBROUTINE NcDef_var_attributes_r8(ncid,var_id,att_name,att_val)
+  SUBROUTINE NcDef_var_attributes_r8(ncid, var_id, att_name, att_val)
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT none
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !!  ncid    : netCDF file id
@@ -462,13 +455,11 @@ CONTAINS
 !
 ! !LOCAL VARIABLES:
     CHARACTER (LEN=512) :: err_msg
-    INTEGER             ::  mylen, ierr
+    INTEGER             :: ierr
 !
-    mylen = 1
-    ierr  = Nf_Put_Att_Double( ncid,      var_id, att_name, &
-                               NF_DOUBLE, mylen,  att_val )
+    ierr  = NF90_Put_Att( ncid, var_id, att_name, att_val )
 
-    IF (ierr.ne.NF_NOERR) THEN
+    IF (ierr.ne.NF90_NOERR) THEN
        err_msg = 'NcDef_var_attributes_r8: can not define attribute : ' // &
             TRIM (att_name)
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
@@ -486,14 +477,12 @@ CONTAINS
 !
 ! !INTERFACE:
 !
-  SUBROUTINE NcDef_var_attributes_i_arr(ncid,var_id,att_name,att_val)
+  SUBROUTINE NcDef_var_attributes_i_arr(ncid, var_id, att_name, att_val)
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT none
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !!  ncid    : netCDF file id
@@ -519,13 +508,11 @@ CONTAINS
 !
 ! !LOCAL VARIABLES:
     CHARACTER (LEN=512) :: err_msg
-    INTEGER             :: mylen, ierr
+    INTEGER             :: ierr
 !
-    mylen = SIZE( att_val )
-    ierr  = Nf_Put_Att_Int( ncid,   var_id, att_name, &
-                            NF_INT, mylen,  att_val )
+    ierr  = NF90_Put_Att( ncid, var_id, att_name, att_val )
 
-    iF (ierr.ne.NF_NOERR) THEN
+    iF (ierr.ne.NF90_NOERR) THEN
        err_msg = 'NcDef_var_attributes_i_arr: can not define attribute : ' &
             // TRIM (att_name)
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
@@ -543,14 +530,12 @@ CONTAINS
 !
 ! !INTERFACE:
 !
-  SUBROUTINE NcDef_var_attributes_r4_arr(ncid,var_id,att_name,att_val)
+  SUBROUTINE NcDef_var_attributes_r4_arr(ncid, var_id, att_name, att_val)
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT none
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !!  ncid    : netCDF file id
@@ -576,13 +561,11 @@ CONTAINS
 !
 ! !LOCAL VARIABLES:
     CHARACTER (LEN=512) :: err_msg
-    INTEGER             :: mylen, ierr
+    INTEGER             :: ierr
 !
-    mylen = SIZE( att_val )
-    ierr  = Nf_Put_Att_Real( ncid,     var_id, att_name, &
-                             NF_FLOAT, mylen,  att_val )
+    ierr  = NF90_Put_Att( ncid, var_id, att_name, att_val )
 
-    IF (ierr.ne.NF_NOERR) THEN
+    IF (ierr.ne.NF90_NOERR) THEN
        err_msg = 'NcDef_var_attributes_r4_arr: can not define attribute : ' &
                     // TRIM (att_name)
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
@@ -600,14 +583,12 @@ CONTAINS
 !
 ! !INTERFACE:
 !
-  SUBROUTINE NcDef_var_attributes_r8_arr(ncid,var_id,att_name,att_val)
+  SUBROUTINE NcDef_var_attributes_r8_arr(ncid, var_id, att_name, att_val)
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT NONE
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !!    ncid    : netCDF file id
@@ -633,13 +614,11 @@ CONTAINS
 !
 ! !LOCAL VARIABLES:
     character (len=512) :: err_msg
-    integer             ::  mylen, ierr
+    integer             :: ierr
 !
-    mylen = size( att_val )
-    ierr  = Nf_Put_Att_Double( ncid,      var_id, att_name, &
-                               NF_DOUBLE, mylen,  att_val )
+    ierr  = NF90_Put_Att( ncid, var_id, att_name, att_val )
 
-    IF (ierr.ne.NF_NOERR) THEN
+    IF (ierr.ne.NF90_NOERR) THEN
        err_msg = 'NcDef_var_attributes_r4_arr: can not define attribute : '&
                      // Trim (att_name)
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
@@ -657,15 +636,12 @@ CONTAINS
 !
 ! !INTERFACE:
 !
-  SUBROUTINE NcDef_glob_attributes_c(ncid,att_name,att_val)
+  SUBROUTINE NcDef_glob_attributes_c(ncid, att_name, att_val)
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT NONE
-!
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !!  ncid    : netCDF file id
@@ -691,12 +667,11 @@ CONTAINS
 !
 ! !LOCAL VARIABLES:
     CHARACTER (LEN=512) :: err_msg
-    INTEGER             ::  mylen, ierr
+    INTEGER             :: ierr
 !
-    mylen = len(att_val)
-    ierr = Nf_Put_Att_Text (ncid, NF_GLOBAL, att_name, mylen, att_val)
+    ierr = NF90_Put_Att(ncid, NF90_GLOBAL, att_name, att_val)
 
-    IF (ierr.ne.NF_NOERR) THEN
+    IF (ierr.ne.NF90_NOERR) THEN
        err_msg = 'NcDef_glob_attributes_c: can not define attribute : ' // &
             TRIM (att_name)
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
@@ -714,15 +689,12 @@ CONTAINS
 !
 ! !INTERFACE:
 !
-  SUBROUTINE NcDef_glob_attributes_i(ncid,att_name,att_val)
+  SUBROUTINE NcDef_glob_attributes_i(ncid, att_name, att_val)
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT none
-!
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !!    ncid    : netCDF file id
@@ -748,13 +720,11 @@ CONTAINS
 !
 ! !LOCAL VARIABLES:
     CHARACTER (LEN=512) :: err_msg
-    INTEGER             :: mylen, ierr
+    INTEGER             :: ierr
 !
-    mylen = 1
-    ierr  = Nf_Put_Att_Int( ncid,   NF_GLOBAL, att_name, &
-                            NF_INT, mylen,     att_val )
+    ierr  = NF90_Put_Att( ncid, NF90_GLOBAL, att_name, att_val )
 
-    IF (ierr.ne.NF_NOERR) THEN
+    IF (ierr.ne.NF90_NOERR) THEN
        err_msg = 'NcDef_glob_attributes_i: can not define attribute : ' // &
             TRIM (att_name)
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
@@ -772,15 +742,12 @@ CONTAINS
 !
 ! !INTERFACE:
 !
-  SUBROUTINE NcDef_glob_attributes_r4(ncid,att_name,att_val)
+  SUBROUTINE NcDef_glob_attributes_r4(ncid, att_name, att_val)
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT NONE
-!
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !!  ncid    : netCDF file id
@@ -806,13 +773,11 @@ CONTAINS
 !
 ! !LOCAL VARIABLES:
     character (len=512) :: err_msg
-    integer             :: mylen, ierr
+    integer             :: ierr
 !
-    mylen = 1
-    ierr  = Nf_Put_Att_Real( ncid,     NF_GLOBAL, att_name, &
-                             NF_FLOAT, mylen,     att_val )
+    ierr  = NF90_Put_Att( ncid, NF90_GLOBAL, att_name, att_val )
 
-    IF (ierr.ne.NF_NOERR) THEN
+    IF (ierr.ne.NF90_NOERR) THEN
        err_msg = 'NcDef_glob_attributes_r4: can not define attribute : ' // &
             TRIM (att_name)
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
@@ -830,15 +795,12 @@ CONTAINS
 !
 ! !INTERFACE:
 !
-  SUBROUTINE NcDef_glob_attributes_r8(ncid,att_name,att_val)
+  SUBROUTINE NcDef_glob_attributes_r8(ncid, att_name, att_val)
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT NONE
-!
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !!    ncid    : netCDF file id
@@ -864,13 +826,11 @@ CONTAINS
 !
 ! !LOCAL VARIABLES:
     character (len=512) :: err_msg
-    integer             :: mylen, ierr
+    integer             :: ierr
 !
-    mylen = 1
-    ierr  = Nf_Put_Att_Double( ncid,     NF_GLOBAL, att_name, &
-                               NF_FLOAT, mylen,     att_val )
+    ierr  = NF90_Put_Att( ncid, NF90_GLOBAL, att_name, att_val )
 
-    IF (ierr.ne.NF_NOERR) THEN
+    IF (ierr.ne.NF90_NOERR) THEN
        err_msg = 'NcDef_glob_attributes_r8: can not define attribute : ' // &
             TRIM (att_name)
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
@@ -888,15 +848,12 @@ CONTAINS
 !
 ! !INTERFACE:
 !
-  SUBROUTINE NcDef_glob_attributes_i_arr(ncid,att_name,att_val)
+  SUBROUTINE NcDef_glob_attributes_i_arr(ncid, att_name, att_val)
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT NONE
-!
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !!    ncid    : netCDF file id
@@ -922,13 +879,11 @@ CONTAINS
 !
 ! !LOCAL VARIABLES:
     CHARACTER (LEN=512) :: err_msg
-    INTEGER             :: mylen, ierr
+    INTEGER             :: ierr
 !
-    mylen = SIZE( att_val )
-    ierr  = Nf_Put_Att_Int( ncid,   NF_GLOBAL, att_name, &
-                            NF_INT, mylen,     att_val )
+    ierr  = NF90_Put_Att( ncid, NF90_GLOBAL, att_name, att_val )
 
-    IF (ierr.ne.NF_NOERR) THEN
+    IF (ierr.ne.NF90_NOERR) THEN
        err_msg = 'NcDef_glob_attributes_i_arr: can not define attribute : ' &
             // Trim (att_name)
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
@@ -950,11 +905,8 @@ CONTAINS
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT none
-!
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !!  ncid    : netCDF file id
@@ -980,13 +932,11 @@ CONTAINS
 !
 ! !LOCAL VARIABLES:
     character (len=512) :: err_msg
-    integer             :: mylen, ierr
+    integer             :: ierr
 !
-    mylen = SIZE( att_val )
-    ierr  = Nf_Put_Att_Real( ncid,     NF_GLOBAL, att_name, &
-                             NF_FLOAT, mylen,     att_val )
+    ierr  = NF90_Put_Att( ncid, NF90_GLOBAL, att_name, att_val )
 
-    IF (ierr.ne.NF_NOERR) THEN
+    IF (ierr.ne.NF90_NOERR) THEN
        err_msg = 'NcDef_glob_attributes_r4_arr: can not define attribute : ' &
               // TRIM (att_name)
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
@@ -1004,15 +954,12 @@ CONTAINS
 !
 ! !INTERFACE:
 !
-  SUBROUTINE NcDef_glob_attributes_r8_arr(ncid,att_name,att_val)
+  SUBROUTINE NcDef_glob_attributes_r8_arr(ncid, att_name, att_val)
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT none
-!
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !!  ncid    : netCDF file id
@@ -1038,13 +985,11 @@ CONTAINS
 !
 ! !LOCAL VARIABLES:
     character (len=512) :: err_msg
-    integer             :: mylen, ierr
+    integer             :: ierr
 !
-    mylen = SIZE( att_val )
-    ierr  = Nf_Put_Att_Double( ncid,     NF_GLOBAL, att_name, &
-                               NF_FLOAT, mylen,     att_val )
+    ierr  = NF90_Put_Att( ncid, NF90_GLOBAL, att_name, att_val )
 
-    IF (ierr.ne.NF_NOERR) THEN
+    IF (ierr.ne.NF90_NOERR) THEN
        err_msg = 'NcDef_glob_attributes_r8_arr: can not define attribute : ' &
             // TRIM (att_name)
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
@@ -1062,19 +1007,17 @@ CONTAINS
 !
 ! !INTERFACE:
 !
-  SUBROUTINE NcSetFill(ncid,ifill,omode)
+  SUBROUTINE NcSetFill(ncid, ifill, omode)
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT NONE
-!
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !
-    INTEGER, INTENT(in) :: ncid, ifill,omode
+    INTEGER, INTENT(IN   ) :: ncid, ifill
+    INTEGER, INTENT(INOUT) :: omode
 !
 ! !DESCRIPTION: Sets fill method.
 !\\
@@ -1090,12 +1033,12 @@ CONTAINS
 !
 ! !LOCAL VARIABLES:
     character (len=512) :: err_msg
-    integer             ::  mylen, ierr
+    integer             :: mylen, ierr
 !
-    ierr = Nf_Set_Fill (ncid, NF_NOFILL, omode)
+    ierr = NF90_Set_Fill(ncid, NF90_NOFILL, omode)
 
-    IF (ierr.ne.NF_NOERR) THEN
-       err_msg = 'Nf_Set_FIll: Error in omode  '
+    IF (ierr.ne.NF90_NOERR) THEN
+       err_msg = 'NF90_Set_FIll: Error in omode  '
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
     END IF
 
@@ -1115,11 +1058,8 @@ CONTAINS
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT NONE
-!
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !
@@ -1141,10 +1081,10 @@ CONTAINS
     CHARACTER (LEN=512) :: err_msg
     INTEGER             ::  ierr
 !
-    ierr = Nf_Enddef (ncid)
+    ierr = NF90_Enddef(ncid)
 
-    IF (ierr.ne.NF_NOERR) THEN
-       err_msg = 'Nf_EndDef: Error in closing netCDF define mode!'
+    IF (ierr.ne.NF90_NOERR) THEN
+       err_msg = 'NF90_EndDef: Error in closing netCDF define mode!'
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
     END IF
 
@@ -1164,11 +1104,8 @@ CONTAINS
 !
 ! !USES:
 !
+    USE netCDF
     USE m_do_err_out
-!
-    IMPLICIT none
-!
-    INCLUDE 'netcdf.inc'
 !
 ! !INPUT PARAMETERS:
 !
@@ -1192,10 +1129,10 @@ CONTAINS
     character (len=512) :: err_msg
     integer             :: ierr
 !
-    ierr = Nf_Redef (ncid)
+    ierr = NF90_Redef (ncid)
 
-    IF (ierr.ne.NF_NOERR) THEN
-       err_msg = 'Nf_ReDef: Error in opening netCDF define mode!'
+    IF (ierr.ne.NF90_NOERR) THEN
+       err_msg = 'NF90_ReDef: Error in opening netCDF define mode!'
        CALL Do_Err_Out (err_msg, .true., 0, 0, 0, 0, 0.0d0, 0.0d0)
     END IF