+++ src/Shared/NcdfUtil/hco_m_netcdf_io_readattr.F90
@@ -11,13 +11,9 @@
 MODULE HCO_m_netcdf_io_readattr
 !
 ! !USES:
-
-  USE m_do_err_out
-
+!
   IMPLICIT NONE
   PRIVATE
-
-  INCLUDE "netcdf.inc"
 !
 ! !PUBLIC MEMBER FUNCTIONS:
 !
@@ -93,6 +89,11 @@ CONTAINS
 !
   SUBROUTINE NcGet_Var_Attr_C( fid, varName, attName, attValue )
 !
+! USES:
+!
+    USE netCDF
+    USE m_do_err_out
+!
 ! !INPUT PARAMETERS:
 !
     INTEGER,          INTENT(IN)  :: fId        ! netCDF file ID
@@ -125,20 +126,20 @@ CONTAINS
     attValue = ''
 
     ! Check if VARNAME is a valid variable
-    status = Nf_Inq_Varid ( fId, varName, vId )
+    status = NF90_Inq_VarId( fId, varName, vId )
 
     ! Exit w/ error message if VARNAME is not valid
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Var_Attr_C: ' // TRIM( varName )        // &
-                 ', '                   // Nf_Strerror( status )
+                 ', '                   // NF90_Strerror( status )
        CALL Do_Err_Out( errMsg, .TRUE., 0, 0, 0, 0, 0.0d0, 0.0d0 )
     ENDIF
 
     !  Get the attribute
-    status = Nf_Get_Att_Text( fId, vId, attName, attValue )
+    status = NF90_Get_Att( fId, vId, attName, attValue )
 
     ! Exit w/ error message if unsuccessful
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Var_Attr_C: cannot read attribute : ' // &
                  TRIM( attName )
        CALL Do_Err_Out( errMsg, .TRUE., 0, 0, 0, 0, 0.0d0, 0.0d0 )
@@ -160,6 +161,11 @@ CONTAINS
 ! !INTERFACE:
 !
   SUBROUTINE NcGet_Var_Attr_I4( fid, varName, attName, attValue )
+!
+! USES:
+!
+    USE netCDF
+    USE m_do_err_out
 !
 ! !INPUT PARAMETERS:
 !
@@ -193,20 +199,20 @@ CONTAINS
     attValue = 0
 
     ! Check if VARNAME is a valid variable
-    status = Nf_Inq_Varid ( fId, varName, vId )
+    status = NF90_Inq_VarId( fId, varName, vId )
 
     ! Exit w/ error message if VARNAME is not valid
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Var_Attr_I4: ' // TRIM( varName )        // &
-                 ', '                   // Nf_Strerror( status )
+                 ', '                   // NF90_Strerror( status )
        CALL Do_Err_Out ( errMsg, .TRUE., 1, fId, 0, 0, 0.0d0, 0.0d0)
     ENDIF
 
     ! Get the attribute
-    status = Nf_Get_Att_Int( fId, vId, attName, attValue )
+    status = NF90_Get_Att( fId, vId, attName, attValue )
 
     ! Exit w/ error message if unsuccessful
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Var_Attr_I4: cannot read attribute : ' // &
                  TRIM( attName )
        CALL Do_Err_Out( errMsg, .TRUE., 0, 0, 0, 0, 0.0d0, 0.0d0 )
@@ -228,6 +234,11 @@ CONTAINS
 ! !INTERFACE:
 !
   SUBROUTINE NcGet_Var_Attr_R4( fid, varName, attName, attValue )
+!
+! USES:
+!
+    USE netCDF
+    USE m_do_err_out
 !
 ! !INPUT PARAMETERS:
 !
@@ -261,20 +272,20 @@ CONTAINS
     attValue = 0e0
 
     ! Check if VARNAME is a valid variable
-    status = Nf_Inq_Varid ( fId, varName, vId )
+    status = NF90_Inq_VarId( fId, varName, vId )
 
     ! Exit w/ error message if VARNAME is not valid
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Var_Attr_R4: ' // TRIM( varName )        // &
-                 ', '                   // Nf_Strerror( status )
+                 ', '                   // NF90_Strerror( status )
        CALL Do_Err_Out ( errMsg, .TRUE., 1, fId, 0, 0, 0.0d0, 0.0d0)
     ENDIF
 
     ! Get the attribute
-    status = Nf_Get_Att_Real( fId, vId, attName, attValue )
+    status = NF90_Get_Att( fId, vId, attName, attValue )
 
     ! Exit w/ error message if unsuccessful
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Var_Attr_R4: cannot read attribute : ' // &
                  TRIM( attName )
        CALL Do_Err_Out( errMsg, .TRUE., 0, 0, 0, 0, 0.0d0, 0.0d0 )
@@ -296,6 +307,11 @@ CONTAINS
 ! !INTERFACE:
 !
   SUBROUTINE NcGet_Var_Attr_R8( fid, varName, attName, attValue )
+!
+! USES:
+!
+    USE netCDF
+    USE m_do_err_out
 !
 ! !INPUT PARAMETERS:
 !
@@ -329,20 +345,20 @@ CONTAINS
     attValue = 0d0
 
     ! Check if VARNAME is a valid variable
-    status = Nf_Inq_Varid ( fId, varName, vId )
+    status = NF90_Inq_VarId( fId, varName, vId )
 
     ! Exit w/ error message if VARNAME is not valid
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Var_Attr_R8: ' // TRIM( varName )        // &
-                 ', '                   // Nf_Strerror( status )
+                 ', '                   // NF90_Strerror( status )
        CALL Do_Err_Out ( errMsg, .TRUE., 1, fId, 0, 0, 0.0d0, 0.0d0)
     ENDIF
 
     ! Get the attribute
-    status = Nf_Get_Att_Double( fId, vId, attName, attValue )
+    status = NF90_Get_Att( fId, vId, attName, attValue )
 
     ! Exit w/ error message if unsuccessful
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Var_Attr_R8: cannot read attribute : ' // &
                  TRIM( attName )
        CALL Do_Err_Out( errMsg, .TRUE., 0, 0, 0, 0, 0.0d0, 0.0d0 )
@@ -364,6 +380,11 @@ CONTAINS
 ! !INTERFACE:
 !
   SUBROUTINE NcGet_Var_Attr_I4_arr( fid, varName, attName, attValue )
+!
+! USES:
+!
+    USE netCDF
+    USE m_do_err_out
 !
 ! !INPUT PARAMETERS:
 !
@@ -397,20 +418,20 @@ CONTAINS
     attValue = 0
 
     ! Check if VARNAME is a valid variable
-    status = Nf_Inq_Varid ( fId, varName, vId )
+    status = NF90_Inq_VarId( fId, varName, vId )
 
     ! Exit w/ error message if VARNAME is not valid
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Var_Attr_I4_arr: ' // TRIM( varName )        // &
-                 ', '                        // Nf_Strerror( status )
+                 ', '                        // NF90_Strerror( status )
        CALL Do_Err_Out ( errMsg, .TRUE., 1, fId, 0, 0, 0.0d0, 0.0d0)
     ENDIF
 
     ! Get the attribute
-    status = Nf_Get_Att_Int( fId, vId, attName, attValue )
+    status = NF90_Get_Att( fId, vId, attName, attValue )
 
     ! Exit w/ error message if unsuccessful
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Var_Attr_I4_arr: cannot read attribute : ' // &
                  TRIM( attName )
        CALL Do_Err_Out( errMsg, .TRUE., 0, 0, 0, 0, 0.0d0, 0.0d0 )
@@ -432,6 +453,11 @@ CONTAINS
 ! !INTERFACE:
 !
   SUBROUTINE NcGet_Var_Attr_R4_arr( fid, varName, attName, attValue )
+!
+! USES:
+!
+    USE netCDF
+    USE m_do_err_out
 !
 ! !INPUT PARAMETERS:
 !
@@ -465,20 +491,20 @@ CONTAINS
     attValue = 0e0
 
     ! Check if VARNAME is a valid variable
-    status = Nf_Inq_Varid ( fId, varName, vId )
+    status = NF90_Inq_VarId( fId, varName, vId )
 
     ! Exit w/ error message if VARNAME is not valid
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Var_Attr_R4_arr: ' // TRIM( varName )        // &
-                 ', '                        // Nf_Strerror( status )
+                 ', '                        // NF90_Strerror( status )
        CALL Do_Err_Out ( errMsg, .TRUE., 1, fId, 0, 0, 0.0d0, 0.0d0)
     ENDIF
 
     ! Get the attribute
-    status = Nf_Get_Att_Real( fId, vId, attName, attValue )
+    status = NF90_Get_Att( fId, vId, attName, attValue )
 
     ! Exit w/ error message if unsuccessful
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Var_Attr_R4_arr: cannot read attribute : ' // &
                  TRIM( attName )
        CALL Do_Err_Out( errMsg, .TRUE., 0, 0, 0, 0, 0.0d0, 0.0d0 )
@@ -500,6 +526,11 @@ CONTAINS
 ! !INTERFACE:
 !
   SUBROUTINE NcGet_Var_Attr_R8_arr( fid, varName, attName, attValue )
+!
+! USES:
+!
+    USE netCDF
+    USE m_do_err_out
 !
 ! !INPUT PARAMETERS:
 !
@@ -533,20 +564,20 @@ CONTAINS
     attValue = 0d0
 
     ! Check if VARNAME is a valid variable
-    status = Nf_Inq_Varid ( fId, varName, vId )
+    status = NF90_Inq_VarId( fId, varName, vId )
 
     ! Exit w/ error message if VARNAME is not valid
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Var_Attr_R8_arr: ' // TRIM( varName )        // &
-                 ', '                        // Nf_Strerror( status )
+                 ', '                        // NF90_Strerror( status )
        CALL Do_Err_Out ( errMsg, .TRUE., 1, fId, 0, 0, 0.0d0, 0.0d0)
     ENDIF
 
     ! Get the attribute
-    status = Nf_Get_Att_Double( fId, vId, attName, attValue )
+    status = NF90_Get_Att( fId, vId, attName, attValue )
 
     ! Exit w/ error message if unsuccessful
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Var_Attr_R8_arr: cannot read attribute : ' // &
                  TRIM( attName )
        CALL Do_Err_Out( errMsg, .TRUE., 0, 0, 0, 0, 0.0d0, 0.0d0 )
@@ -568,6 +599,11 @@ CONTAINS
 ! !INTERFACE:
 !
   SUBROUTINE NcGet_Glob_Attr_C( fid, attName, attValue )
+!
+! USES:
+!
+    USE netCDF
+    USE m_do_err_out
 !
 ! !INPUT PARAMETERS:
 !
@@ -600,10 +636,10 @@ CONTAINS
     attValue = ''
 
     ! Get the attribute
-    status = Nf_Get_Att_Text( fId, NF_GLOBAL, attName, attValue )
+    status = NF90_Get_Att( fId, NF90_GLOBAL, attName, attValue )
 
     ! Exit w/ error message if unsuccessful
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Glob_Attr_C: cannot read attribute : ' // &
                  TRIM( attName )
        CALL Do_Err_Out( errMsg, .TRUE., 0, 0, 0, 0, 0.0d0, 0.0d0 )
@@ -625,6 +661,11 @@ CONTAINS
 ! !INTERFACE:
 !
   SUBROUTINE NcGet_Glob_Attr_I4( fid, attName, attValue )
+!
+! USES:
+!
+    USE netCDF
+    USE m_do_err_out
 !
 ! !INPUT PARAMETERS:
 !
@@ -657,10 +698,10 @@ CONTAINS
     attValue = 0
 
     ! Get the attribute
-    status = Nf_Get_Att_Int( fId, NF_GLOBAL, attName, attValue )
+    status = NF90_Get_Att( fId, NF90_GLOBAL, attName, attValue )
 
     ! Exit w/ error message if unsuccessful
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Glob_Attr_I4: cannot read attribute : ' // &
                  TRIM( attName )
        CALL Do_Err_Out( errMsg, .TRUE., 0, 0, 0, 0, 0.0d0, 0.0d0 )
@@ -682,6 +723,11 @@ CONTAINS
 ! !INTERFACE:
 !
   SUBROUTINE NcGet_Glob_Attr_R4( fid, attName, attValue )
+!
+! USES:
+!
+    USE netCDF
+    USE m_do_err_out
 !
 ! !INPUT PARAMETERS:
 !
@@ -714,10 +760,10 @@ CONTAINS
     attValue = 0e0
 
     ! Get the attribute
-    status = Nf_Get_Att_Real( fId, NF_GLOBAL, attName, attValue )
+    status = NF90_Get_Att( fId, NF90_GLOBAL, attName, attValue )
 
     ! Exit w/ error message if unsuccessful
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Glob_Attr_R4: cannot read attribute : ' // &
                  TRIM( attName )
        CALL Do_Err_Out( errMsg, .TRUE., 0, 0, 0, 0, 0.0d0, 0.0d0 )
@@ -739,6 +785,11 @@ CONTAINS
 ! !INTERFACE:
 !
   SUBROUTINE NcGet_Glob_Attr_R8( fid, attName, attValue )
+!
+! USES:
+!
+    USE netCDF
+    USE m_do_err_out
 !
 ! !INPUT PARAMETERS:
 !
@@ -771,10 +822,10 @@ CONTAINS
     attValue = 0d0
 
     ! Get the attribute
-    status = Nf_Get_Att_Double( fId, NF_GLOBAL, attName, attValue )
+    status = NF90_Get_Att( fId, NF90_GLOBAL, attName, attValue )
 
     ! Exit w/ error message if unsuccessful
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Glob_Attr_R8: cannot read attribute : ' // &
                  TRIM( attName )
        CALL Do_Err_Out( errMsg, .TRUE., 0, 0, 0, 0, 0.0d0, 0.0d0 )
@@ -796,6 +847,11 @@ CONTAINS
 ! !INTERFACE:
 !
   SUBROUTINE NcGet_Glob_Attr_I4_arr( fid, attName, attValue )
+!
+! USES:
+!
+    USE netCDF
+    USE m_do_err_out
 !
 ! !INPUT PARAMETERS:
 !
@@ -828,10 +884,10 @@ CONTAINS
     attValue = 0
 
     ! Get the attribute
-    status = Nf_Get_Att_Int( fId, NF_GLOBAL, attName, attValue )
+    status = NF90_Get_Att( fId, NF90_GLOBAL, attName, attValue )
 
     ! Exit w/ error message if unsuccessful
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Glob_Attr_I4_arr: cannot read attribute : ' // &
                  TRIM( attName )
        CALL Do_Err_Out( errMsg, .TRUE., 0, 0, 0, 0, 0.0d0, 0.0d0 )
@@ -853,6 +909,11 @@ CONTAINS
 ! !INTERFACE:
 !
   SUBROUTINE NcGet_Glob_Attr_R4_arr( fid, attName, attValue )
+!
+! USES:
+!
+    USE netCDF
+    USE m_do_err_out
 !
 ! !INPUT PARAMETERS:
 !
@@ -885,10 +946,10 @@ CONTAINS
     attValue = 0e0
 
     ! Get the attribute
-    status = Nf_Get_Att_Real( fId, NF_GLOBAL, attName, attValue )
+    status = NF90_Get_Att( fId, NF90_GLOBAL, attName, attValue )
 
     ! Exit w/ error message if unsuccessful
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Glob_Attr_R4_arr: cannot read attribute : ' // &
                  TRIM( attName )
        CALL Do_Err_Out( errMsg, .TRUE., 0, 0, 0, 0, 0.0d0, 0.0d0 )
@@ -910,6 +971,11 @@ CONTAINS
 ! !INTERFACE:
 !
   SUBROUTINE NcGet_Glob_Attr_R8_arr( fid, attName, attValue )
+!
+! USES:
+!
+    USE netCDF
+    USE m_do_err_out
 !
 ! !INPUT PARAMETERS:
 !
@@ -942,10 +1008,10 @@ CONTAINS
     attValue = 0d0
 
     ! Get the attribute
-    status = Nf_Get_Att_Double( fId, NF_GLOBAL, attName, attValue )
+    status = NF90_Get_Att( fId, NF90_GLOBAL, attName, attValue )
 
     ! Exit w/ error message if unsuccessful
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        errMsg = 'In NcGet_Glob_Attr_R8_arr: cannot read attribute : ' // &
                  TRIM( attName )
        CALL Do_Err_Out( errMsg, .TRUE., 0, 0, 0, 0, 0.0d0, 0.0d0 )
@@ -969,6 +1035,11 @@ CONTAINS
 ! !INTERFACE:
 !
   SUBROUTINE NcGet_Var_Attr_C_nostop( fId, varName, attName, attValue, RC )
+!
+! USES:
+!
+    USE netCDF
+    USE m_do_err_out
 !
 ! !INPUT PARAMETERS:
 !
@@ -1003,19 +1074,19 @@ CONTAINS
     attValue = ''
 
     ! Check if VARNAME is a valid variable
-    status = Nf_Inq_Varid ( fId, varName, vId )
+    status = NF90_Inq_VarId( fId, varName, vId )
 
     ! Exit w/ error message if VARNAME is not valid
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        RC = status
        RETURN
     ENDIF
 
     !  Get the attribute
-    status = Nf_Get_Att_Text( fId, vId, attName, attValue )
+    status = NF90_Get_Att( fId, vId, attName, attValue )
 
     ! Exit w/ error message if unsuccessful
-    IF ( status /= NF_NOERR ) THEN
+    IF ( status /= NF90_NOERR ) THEN
        RC = status
        RETURN
     ENDIF