+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- NetCDF routines in `src/Shared/NcdfUtil` now use the Fortran-90 API
+
 ## [Unreleased 3.7.1] - TBD
 ### Changed
 - Updated version numbers to 3.7.1